feat: login
* feat: user login routes and logic

* feat: login backend

* feat: error message when unauthorized

* fix user

* feat: frontend login

* feat: full login

* feat: password

* remove unused imports

* feat: tasks per user

* feat: do not send hashed password

* refactor task mutation

+++ todoapp_backend/src/routes/tasks.rs
@@ -1,26 +1,36 @@
 use rocket::{Route, serde::json::Json};
 
-use crate::{Database, models::tasks::{Task, TaskInsert}};
+use crate::{Database, core::auth::Auth, models::tasks::{Task, TaskInsert, TaskUpdate}};
 
 
 #[get("/")]
-pub async fn list_tasks(conn: Database) -> Json<Vec<Task>> {
-    Json(Task::get_all(&conn).await)
+pub async fn list_tasks(conn: Database, auth: Auth) -> Json<Vec<Task>> {
+    Json(Task::get_all(&conn, auth.user.id).await)
 }
 
 #[post("/", data = "<todo>")]
-pub async fn create_task(conn: Database, todo: Json<TaskInsert>) -> Json<Task> {
-    Json(todo.save(&conn).await)
+pub async fn create_task(conn: Database, todo: Json<TaskInsert>, auth: Auth) -> Json<Task> {
+    Json(todo.into_inner().save(&conn, auth.user.id).await)
 }
 
 #[put("/", data = "<todo>")]
-pub async fn update_task(conn: Database, todo: Json<Task>) {
-    todo.update(&conn).await
+pub async fn update_task(conn: Database, todo: Json<TaskUpdate>, auth: Auth) -> Result<(), String> {
+    let task = Task::get(&conn, todo.id).await.ok_or("Task not found")?;
+    if task.user_id != auth.user.id {
+        return Err("The task does not belong to this user".into());
+    }
+    todo.update(&conn).await;
+    Ok(())
 }
 
 #[delete("/<id>")]
-pub async fn delete_task(conn: Database, id: i32) {
-    Task::delete(&conn, id).await
+pub async fn delete_task(conn: Database, id: i32, auth: Auth) -> Result<(), String> {
+    let task = Task::get(&conn, id).await.ok_or("Task not found")?;
+    if task.user_id != auth.user.id {
+        return Err("The task does not belong to this user".into());
+    }
+    Task::delete(&conn, id).await;
+    Ok(())
 }
 
 pub fn get_routes() -> Vec<Route> {
